requirements.txt
APScheduler==3.7.0
bcrypt==3.2.0
blinker==1.4
cachelib==0.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
decorator==5.0.9
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
infinity==1.5
intervals==0.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.20
SQLAlchemy-Utils==0.37.8
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
